a Lot of changes (Pending: solve the question where to put the activation function and also start using the pointer wrappers)

+++ include/libdl/L2LossLayer.h
@@ -35,7 +35,7 @@ void L2LossLayer::ForwardPass()
         //L2
         mGradientHelper = (*mInputPtr) - mLabels;
         // Loss divided by number of examples;
-        mLoss = (0.5/vOutputNum) * mGradientHelper.rowwise().squaredNorm().sum();
+        mLoss = (0.5/static_cast<double>(vOutputNum)) * mGradientHelper.rowwise().squaredNorm().sum();
     }
     else
     {
